warning
svn:sync-xref-src-repo-rev: r86102

+++ src/VBox/Main/src-server/HostImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: HostImpl.cpp 46330 2013-05-30 12:41:48Z noreply@oracle.com $ */
+/* $Id: HostImpl.cpp 46331 2013-05-30 12:56:52Z noreply@oracle.com $ */
 /** @file
  * VirtualBox COM class implementation: Host
  */
@@ -1045,7 +1045,7 @@ STDMETHODIMP Host::COMGETTER(MemorySize)(ULONG *aSize)
     int rc = RTSystemQueryTotalRam(&cb);
     if (RT_FAILURE(rc))
         return E_FAIL;
-    *aSize = cb / _1M;
+    *aSize = (ULONG)(cb / _1M);
     return S_OK;
 }
 
@@ -1064,7 +1064,7 @@ STDMETHODIMP Host::COMGETTER(MemoryAvailable)(ULONG *aAvailable)
     int rc = RTSystemQueryAvailableRam(&cb);
     if (RT_FAILURE(rc))
         return E_FAIL;
-    *aAvailable = cb / _1M;
+    *aAvailable = (ULONG)(cb / _1M);
     return S_OK;
 }
